+++ src/infra/ws-backend.ts
@@ -1,5 +1,6 @@
 import { Construct } from "constructs";
 import * as apiGw2 from "@aws-cdk/aws-apigatewayv2-alpha";
+import * as apigw from "aws-cdk-lib/aws-apigatewayv2";
 import { WebSocketLambdaIntegration } from "@aws-cdk/aws-apigatewayv2-integrations-alpha";
 import * as lambda from "aws-cdk-lib/aws-lambda";
 import * as dynamodb from "aws-cdk-lib/aws-dynamodb";
@@ -13,25 +14,30 @@ export type WSBackendProps = {
   name: string;
   region: string;
   account: string;
+  lambdaTimeout: number;
   certificate: acm.ICertificate;
   domainName: string;
 };
 
 export class WSBackend extends Construct {
   public domain: apiGw2.DomainName;
+  public lambdaFunction: lambda.IFunction;
+  public dynamoTable: dynamodb.ITable;
+  public api: apiGw2.WebSocketApi;
 
   constructor(scope: Construct, id: string, props: WSBackendProps) {
     super(scope, id);
-    const { name, region, account, certificate, domainName } = props;
+    const { name, region, account, certificate, domainName, lambdaTimeout } =
+      props;
 
     // Create dynamo db table to track the sessions
-    const dynamoTable = new dynamodb.Table(this, `${name}-sessions`, {
+    this.dynamoTable = new dynamodb.Table(this, `${name}-sessions`, {
       tableName: `${name}-sessions`,
       partitionKey: { name: "sessionId", type: dynamodb.AttributeType.STRING },
       sortKey: { name: "connectionId", type: dynamodb.AttributeType.STRING },
       readCapacity: 10,
       writeCapacity: 10,
-      removalPolicy: RemovalPolicy.SNAPSHOT,
+      removalPolicy: RemovalPolicy.RETAIN,
       encryption: dynamodb.TableEncryption.AWS_MANAGED,
     });
 
@@ -44,42 +50,43 @@ export class WSBackend extends Construct {
     });
 
     // lambda to deal with the requests
-    const lambdaHandler = new lambda.Function(this, `${name}-lambda`, {
+    this.lambdaFunction = new lambda.Function(this, `${name}-lambda`, {
       functionName: `${name}-lambda`,
       runtime: lambda.Runtime.NODEJS_14_X,
       code: lambda.Code.fromAsset("dist/lambda"),
       handler: "dist/lambda/handler.lambdaHandler",
+      timeout: Duration.seconds(lambdaTimeout),
       logRetention: 30,
       environment: {
-        DYNAMO_TABLE_NAME: dynamoTable.tableName,
+        DYNAMO_TABLE_NAME: this.dynamoTable.tableName,
       },
       reservedConcurrentExecutions: 20,
       deadLetterQueue: dlq,
     });
 
     // allow the lambda to read/write/query the sessions dynamo table
-    dynamoTable.grantReadWriteData(lambdaHandler);
+    this.dynamoTable.grantReadWriteData(this.lambdaFunction);
 
     // Create websocket api gateway
-    const api = new apiGw2.WebSocketApi(this, `${name}-api`, {
+    this.api = new apiGw2.WebSocketApi(this, `${name}-api`, {
       apiName: `${name}-api`,
       description: `WebSocket api for ${name}`,
       connectRouteOptions: {
         integration: new WebSocketLambdaIntegration(
           "handle-connect",
-          lambdaHandler
+          this.lambdaFunction
         ),
       },
       disconnectRouteOptions: {
         integration: new WebSocketLambdaIntegration(
           "handle-disconnect",
-          lambdaHandler
+          this.lambdaFunction
         ),
       },
       defaultRouteOptions: {
         integration: new WebSocketLambdaIntegration(
           "handle-default",
-          lambdaHandler
+          this.lambdaFunction
         ),
       },
     });
@@ -92,34 +99,46 @@ export class WSBackend extends Construct {
     });
 
     // Create the stage and deploy the API
-    const stage = new apiGw2.WebSocketStage(this, `${name}-api-stage`, {
+    const cfnStage = new apigw.CfnStage(this, `${name}-api-stage`, {
       stageName: "live",
       autoDeploy: true,
-      webSocketApi: api,
-      throttle: {
-        rateLimit: 100,
-        burstLimit: 100,
+      apiId: this.api.apiId,
+      defaultRouteSettings: {
+        throttlingRateLimit: 100,
+        throttlingBurstLimit: 100,
+        detailedMetricsEnabled: true,
       },
     });
 
+    const stage = apiGw2.WebSocketStage.fromWebSocketStageAttributes(
+      this,
+      `api-stage-ws`,
+      {
+        api: this.api,
+        stageName: cfnStage.stageName,
+      }
+    );
+
     // link custom domain to the api gw
     new apiGw2.ApiMapping(this, `${name}-custom-domain-mapping`, {
-      api: api,
+      api: this.api,
       domainName: this.domain,
       stage,
     });
 
-    api.grantManageConnections(lambdaHandler);
+    this.api.grantManageConnections(this.lambdaFunction);
     // Create policy to allow Lambda function to use @connections API of API Gateway
     const allowConnectionManagementOnApiGatewayPolicy = new PolicyStatement({
       effect: Effect.ALLOW,
       resources: [
-        `arn:aws:execute-api:${region}:${account}:${api.apiId}/*/*/*/@connections/*`,
+        `arn:aws:execute-api:${region}:${account}:${this.api.apiId}/*/*/*/@connections/*`,
       ],
       actions: ["execute-api:Invoke"],
     });
 
     // Attach custom policy to Lambda function
-    lambdaHandler.addToRolePolicy(allowConnectionManagementOnApiGatewayPolicy);
+    this.lambdaFunction.addToRolePolicy(
+      allowConnectionManagementOnApiGatewayPolicy
+    );
   }
 }