+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.5.1](https://github.com/andrewinci/free-retro/compare/v1.5.0...v1.5.1) (2022-08-14)
+
+
+### Bug Fixes
+
+* Fix unsupported dynamo removalPolicy ([483c3c1](https://github.com/andrewinci/free-retro/commit/483c3c10f49e718f15aeddbe145c1ca30ed29aad))
+
+# [1.5.0](https://github.com/andrewinci/free-retro/compare/v1.4.0...v1.5.0) (2022-08-14)
+
+
+### Features
+
+* Add cloudwatch monitors ([2f9adde](https://github.com/andrewinci/free-retro/commit/2f9adde528d1e98cfb4e72f78cffe2eebf1b5cd4))
+
 # [1.4.0](https://github.com/andrewinci/free-retro/compare/v1.3.0...v1.4.0) (2022-08-13)